# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TernakLomba API Documentation",
        "description": "The **TernakLomba API** is a robust and scalable solution for managing competitive events, participants, scores, and results. \n *   It is designed to support a wide range of competitions, from small local events to large-scale tournaments. \n *  Key features include:\n *  - **Competition Management**: Create, update, and manage competitions with details such as name, description, start/end dates, and status.\n *  - **Participant Management**: Register and manage participants, including their details and competition assignments.\n *  - **Score Management**: Submit and manage scores for participants, with support for multiple judges and scoring criteria.\n *  - **Results and Leaderboards**: Automatically calculate and display competition results, including rankings and leaderboards.\n *  - **Judging Criteria**: Define custom judging criteria for competitions to ensure fair and transparent evaluations.\n *  - **Notifications**: Send real-time notifications to participants and judges about competition updates, scores, and results.\n *  This API is ideal for event organizers, educational institutions, and organizations hosting competitive events.",
        "contact": {
            "name": "the developer",
            "email": "support@ternaklomba.com"
        },
        "license": {
            "name": "Apache 2.0",
...
        },
        "version": "1.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to get Authentication Token",
                "description": "Mengambil Token User",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username / Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil login",
                                    "data": {
                                        "token": "1-adsBASDMzxckopasdkpwqkiqwje"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register to get Authentication Token",
                "description": "Register user",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "username",
                        "in": "query",
                        "description": "Username",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "user@gmail.com"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password"
                        }
                    },
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "John"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Doe"
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Gender",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "male",
                                "female"
                            ]
                        },
                        "example": "male"
                    },
                    {
                        "name": "born_date",
                        "in": "query",
                        "description": "Born date",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        },
                        "example": "2025-03-20"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "success": true,
                                    "status_code": 200,
                                    "message": "Berhasil register",
                                    "data": {
                                        "token": "1-adsBASDMzxckopasdkpwqkiqwje"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
...
}